+++ packages/use/src/useRequest/index.ts
@@ -1,7 +1,7 @@
 import type { Promisable } from 'type-fest';
 import type { MaybeRefOrGetter } from 'vue-demi';
-import { createEventHook, toReactive, toValue } from '@vueuse/core';
-import { computed, isVue2, ref, shallowRef } from 'vue-demi';
+import { createEventHook, toValue } from '@vueuse/core';
+import { reactive, ref, shallowRef } from 'vue-demi';
 
 export interface UseRequestOptions {
   /**
@@ -172,22 +172,10 @@ export function useRequest<
     onFinally: finallyEvent.on,
   };
 
-  const reactive = toReactive(
-    computed(() => ({
-      ...common,
-      /** 服务器响应 */
-      response: (isVue2 ? shallowRef(response.value) : computed(() => shallowRef(response.value))) as unknown as typeof response,
-      /** 服务器响应数据 */
-      data: (isVue2 ? shallowRef(data.value) : computed(() => shallowRef(data.value))) as unknown as typeof data,
-      /** 服务器返回的错误 */
-      error: (isVue2 ? shallowRef(error.value) : computed(() => shallowRef(error.value))) as typeof error,
-    })),
-  );
-
   return {
     ...common,
     /** 方法的响应式代理返回值 */
-    reactive,
+    reactive: reactive(common),
   };
 }